adjust DJ resource limits. Fix git sync to handle ssh. Fix IG chart to use git init container

+++ docker/git/sync.sh
@@ -1,6 +1,5 @@
-#!/usr/bin/env sh
-# Sync git configuration upstream
-
+#!/usr/bin/env bash
+# Sync git configuration upstream.
 
 # Top level directory where git projects will be cloned to.
 GIT_ROOT=${GIT_ROOT:=/git}
@@ -13,10 +12,6 @@ GIT_BRANCH=${GIT_BRANCH:-master}5
 git="git"
 
 export GIT_SSH_COMMAND="ssh -q -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -i /etc/git-secret/ssh"
-
-# IDM may be coming up, and we might not be ready
-# Move idm to init container...
-sleep 30
 
 # We don't know the name of the git repo that was cloned, but there should only be a single config directory under the GIT_ROOT.
 cd ${GIT_ROOT}/*
@@ -33,6 +28,11 @@ do
     t=`date`
     $git add .
     $git commit -a -m "autosave at $t"
-    # We use -f to force the upstream push. Revisit this.
-    $git push --set-upstream origin autosave -f
+    # Push is only supported if the upstream repo starts with ssh://
+    # The expression below is bash, not always the same as sh.
+    if [[ ${GIT_REPO} == ssh* ]]; 
+    then
+        # We use -f to force the upstream push. Revisit this.
+        $git push --set-upstream origin autosave -f
+    fi
 done
